+++ Frontend/src/pages/WhoWeAre.jsx
@@ -1,10 +1,13 @@
 import React from "react";
 const WhoWeAre = () => {
   return (
-    <div className="container mx-auto p-6 text-center">
-      <h1 className="text-3xl font-bold">Who We Are</h1>
-      <p className="text-gray-600 mt-4">Learn more about our mission and team.</p>
-    </div>
+    <section ref={aboutRef} className="p-12 text-center bg-gray-100">
+        <h2 className="text-4xl font-bold">Who We Are</h2>
+        <p className="mt-4 text-gray-700">Learn more about our mission and team.</p>
+        <button className="mt-4 bg-blue-500 text-white px-6 py-2 rounded hover:bg-blue-600 transition" onClick={() => scrollToSection(managementRef)}>
+          Management Team
+        </button>
+      </section>
   );
 };
 export default WhoWeAre;
